各種バリデーション

+++ db/migrate/20250105053344_change_email_constraints_on_users.rb
@@ -1,6 +1,6 @@
 class ChangeEmailConstraintsOnUsers < ActiveRecord::Migration[7.2]
   def change
     change_column_null :users, :email, true  # nullを許可
-    remove_index :users, :email  # emailのユニークインデックスを削除
+    remove_index :users, :email if index_exists?(:users, :email)  # emailのユニークインデックスを削除
   end
 end
